+++ weather-api/package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "axios": "^1.5.1",
     "express": "^4.18.2",
+    "node-fetch": "^2.7.0",
     "socket.io": "^4.7.2"
   },
   "name": "weather-app-api",
